Fixed `price` value to use `Decimal` instead
rather than `float` 'cause

- https://docs.python.org/3/library/decimal.html
- https://stackoverflow.com/questions/59904070/how-does-using-decimal-for-money-avoid-the-floating-point-problems-in-python

+++ systems/new_programming/python/Object-Oriented-Programming-with-Python/02_OOP_Python.py
@@ -1,5 +1,6 @@
 import csv
 
+from decimal import Decimal
 from pprint import pprint
 
 
@@ -8,7 +9,7 @@ class Item:
 
     all = []
 
-    def __init__(self, name: str, price: float, quantity: int = 0):
+    def __init__(self, name: str, price: Decimal, quantity: int = 0):
         # in order, to type checking work you need additional apa sih, pokoknya
         # yang ngecek error gitu, soalnya, python nerima nerima aja.
 
@@ -48,7 +49,7 @@ class Item:
             # print(item)
             Item(
                 name=item.get("name"),
-                price=float(
+                price=Decimal(
                     item.get("price")
                 ),  # soalnya di detect nya str, terus float soalnya bisa koma koma
                 quantity=int(item.get("quantity")),
